localization_server revised
Added an argument for the AMCL config file. This is too many arguments.
Have to figure out how to select config files based on the value of the
use_sim_time argument value. For path_planner_server that will be
crucial since there are too many config files to create separate
arguments for.